[5.5] Fix all resources docblocks (#20878)
* Many docblock fixes for resources

* Use Fixtures instead
