+++ src/Illuminate/Http/Resources/Json/PaginatedResourceResponse.php
@@ -10,7 +10,7 @@ class PaginatedResourceResponse extends ResourceResponse
      * Create an HTTP response that represents the object.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Http\JsonResponse
      */
     public function toResponse($request)
     {
