+++ tests/Integration/Http/Fixtures/CommentCollection.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Illuminate\Tests\Integration\Http\Fixtures;
+
+use Illuminate\Http\Resources\Json\ResourceCollection;
+
+class CommentCollection extends ResourceCollection
+{
+    //
+}
