+++ tests/Integration/Http/Fixtures/EmptyPostResource.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Illuminate\Tests\Integration\Http\Fixtures;
+
+class EmptyPostResource extends PostResource
+{
+    //
+}
