+++ tests/Integration/Http/Fixtures/PostResource.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace Illuminate\Tests\Integration\Http\Fixtures;
+
+use Illuminate\Http\Resources\Json\Resource;
+
+class PostResource extends Resource
+{
+    public function toArray($request)
+    {
+        return ['id' => $this->id, 'title' => $this->title, 'custom' => true];
+    }
+
+    public function withResponse($request, $response)
+    {
+        $response->header('X-Resource', 'True');
+    }
+}
