+++ tests/Integration/Http/Fixtures/PostResourceWithOptionalPivotRelationship.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace Illuminate\Tests\Integration\Http\Fixtures;
+
+class PostResourceWithOptionalPivotRelationship extends PostResource
+{
+    public function toArray($request)
+    {
+        return [
+            'id' => $this->id,
+            'subscription' => $this->whenPivotLoaded(Subscription::class, function () {
+                return [
+                    'foo' => 'bar',
+                ];
+            }),
+        ];
+    }
+}
