+++ tests/Integration/Http/Fixtures/PostResourceWithoutWrap.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Illuminate\Tests\Integration\Http\Fixtures;
+
+class PostResourceWithoutWrap extends PostResource
+{
+    public static $wrap = null;
+}
