+++ tests/Integration/Http/Fixtures/Subscription.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Illuminate\Tests\Integration\Http\Fixtures;
+
+class Subscription
+{
+    //
+}
